Make cats belong to specific users
* Clean up links to edit & delete cats
* Forbid editing cats & comments you didn't make
* Make a nice sign in/up page
* Get the specs working with new sign in/up flow

+++ app/controllers/comments_controller.rb
@@ -17,12 +17,16 @@ class CommentsController < ApplicationController
 
   def update
     comment = Comment.find params[:id]
+    head(:unauthorized) and return if current_user != comment.user
+
     comment.update_attributes comment_params
     @comments = comment.cat.comments
   end
 
   def destroy
     comment = Comment.find params[:id]
+    head(:unauthorized) and return if current_user != comment.user
+
     comment.destroy
     @comments = comment.cat.comments
   end
